:recycle: - Break out Build -> Build + Clean + Types
