+++ src/build.rs
@@ -1,100 +1,23 @@
 use crate::bsconfig;
 use crate::bsconfig::OneOrMore;
+use crate::build_types::*;
+use crate::clean;
 use crate::helpers;
-use crate::helpers::get_bs_build_path;
-use crate::helpers::get_package_path;
+use crate::helpers::emojis::*;
 use crate::package_tree;
-use crate::package_tree::Package;
 use ahash::{AHashMap, AHashSet};
-use console::{style, Emoji};
-use indicatif::ProgressBar;
-use indicatif::ProgressStyle;
-use log::Level::Info;
-use log::{debug, error};
-use log::{info, log_enabled};
+use console::style;
+use indicatif::{ProgressBar, ProgressStyle};
+use log::{debug, error, info, log_enabled, Level::Info};
 use rayon::prelude::*;
 use std::fs;
-use std::io::stdout;
-use std::io::Write;
-use std::io::{self, BufRead};
+use std::io::{stdout, Write};
 use std::path::{Path, PathBuf};
 use std::process::Command;
 use std::time::Instant;
-use std::time::SystemTime;
 
-static TREE: Emoji<'_, '_> = Emoji("🌴 ", "");
-static SWEEP: Emoji<'_, '_> = Emoji("🧹 ", "");
-static LOOKING_GLASS: Emoji<'_, '_> = Emoji("🔍 ", "");
-static CODE: Emoji<'_, '_> = Emoji("🟰  ", "");
-static SWORDS: Emoji<'_, '_> = Emoji("⚔️  ", "");
-static CHECKMARK: Emoji<'_, '_> = Emoji("️✅  ", "");
-static CROSS: Emoji<'_, '_> = Emoji("️🛑  ", "");
-static LINE_CLEAR: &str = "\x1b[2K";
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum ParseState {
-    Pending,
-    ParseError,
-    Warning,
-    Success,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum CompileState {
-    Pending,
-    Error,
-    Warning,
-    Success,
-}
-#[derive(Debug, Clone, PartialEq)]
-pub struct Interface {
-    path: String,
-    parse_state: ParseState,
-    compile_state: CompileState,
-    last_modified: SystemTime,
-    dirty: bool,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct Implementation {
-    path: String,
-    parse_state: ParseState,
-    compile_state: CompileState,
-    last_modified: SystemTime,
-    dirty: bool,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct SourceFile {
-    implementation: Implementation,
-    interface: Option<Interface>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct MlMap {
-    dirty: bool,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum SourceType {
-    SourceFile(SourceFile),
-    MlMap(MlMap),
-}
-
-#[derive(Debug, Clone)]
-pub struct Module {
-    pub source_type: SourceType,
-    pub deps: AHashSet<String>,
-    pub package: package_tree::Package,
-}
-
-fn read_lines(filename: String) -> io::Result<io::Lines<io::BufReader<fs::File>>> {
-    let file = fs::File::open(filename)?;
-    Ok(io::BufReader::new(file).lines())
-}
-
-fn get_res_path_from_ast(ast_file: &str) -> Option<String> {
-    if let Ok(lines) = read_lines(ast_file.to_string()) {
+pub fn get_res_path_from_ast(ast_file: &str) -> Option<String> {
+    if let Ok(lines) = helpers::read_lines(ast_file.to_string()) {
         // we skip the first line with is some null characters
         // the following lines in the AST are the dependency modules
         // we stop when we hit a line that starts with a "/", this is the path of the file.
@@ -114,231 +37,11 @@ fn get_res_path_from_ast(ast_file: &str) -> Option<String> {
     return None;
 }
 
-fn remove_asts(source_file: &str, package_name: &str, namespace: &Option<String>, root_path: &str) {
-    let _ = std::fs::remove_file(helpers::get_compiler_asset(
-        source_file,
-        package_name,
-        namespace,
-        root_path,
-        "ast",
-    ));
-    let _ = std::fs::remove_file(helpers::get_compiler_asset(
-        source_file,
-        package_name,
-        namespace,
-        root_path,
-        "iast",
-    ));
-}
-
-fn get_interface<'a>(module: &'a Module) -> &'a Option<Interface> {
+pub fn get_interface<'a>(module: &'a Module) -> &'a Option<Interface> {
     match &module.source_type {
         SourceType::SourceFile(source_file) => &source_file.interface,
         _ => &None,
     }
-}
-
-fn remove_compile_assets(
-    source_file: &str,
-    package_name: &str,
-    namespace: &Option<String>,
-    root_path: &str,
-) {
-    let _ = std::fs::remove_file(helpers::change_extension(source_file, "mjs"));
-    // optimization
-    // only issue cmti if htere is an interfacce file
-    for extension in &["cmj", "cmi", "cmt", "cmti"] {
-        let _ = std::fs::remove_file(helpers::get_compiler_asset(
-            source_file,
-            package_name,
-            namespace,
-            root_path,
-            extension,
-        ));
-        let _ = std::fs::remove_file(helpers::get_bs_compiler_asset(
-            source_file,
-            package_name,
-            namespace,
-            root_path,
-            extension,
-        ));
-    }
-}
-
-pub fn cleanup_previous_build(
-    packages: &AHashMap<String, Package>,
-    all_modules: &mut AHashMap<String, Module>,
-    root_path: &str,
-) -> (usize, usize, AHashSet<String>) {
-    let mut ast_modules: AHashMap<String, (String, String, Option<String>, SystemTime, String)> =
-        AHashMap::new();
-    let mut ast_rescript_file_locations = AHashSet::new();
-
-    let mut rescript_file_locations = all_modules
-        .values()
-        .filter_map(|module| match &module.source_type {
-            SourceType::SourceFile(source_file) => {
-                Some(source_file.implementation.path.to_string())
-            }
-            _ => None,
-        })
-        .collect::<AHashSet<String>>();
-
-    rescript_file_locations.extend(
-        all_modules
-            .values()
-            .filter_map(|module| {
-                get_interface(module)
-                    .as_ref()
-                    .map(|interface| interface.path.to_string())
-            })
-            .collect::<AHashSet<String>>(),
-    );
-
-    // scan all ast files in all packages
-    for package in packages.values() {
-        let read_dir = fs::read_dir(std::path::Path::new(&helpers::get_build_path(
-            root_path,
-            &package.name,
-        )))
-        .unwrap();
-
-        for entry in read_dir {
-            match entry {
-                Ok(entry) => {
-                    let path = entry.path();
-                    let extension = path.extension().and_then(|e| e.to_str());
-                    match extension {
-                        Some(ext) => match ext {
-                            "iast" | "ast" => {
-                                let module_name = helpers::file_path_to_module_name(
-                                    path.to_str().unwrap(),
-                                    &package.namespace,
-                                );
-
-                                let ast_file_path = path.to_str().unwrap().to_owned();
-                                let res_file_path = get_res_path_from_ast(&ast_file_path);
-                                match res_file_path {
-                                    Some(res_file_path) => {
-                                        let _ = ast_modules.insert(
-                                            res_file_path.to_owned(),
-                                            (
-                                                module_name,
-                                                package.name.to_owned(),
-                                                package.namespace.to_owned(),
-                                                entry.metadata().unwrap().modified().unwrap(),
-                                                ast_file_path,
-                                            ),
-                                        );
-                                        let _ = ast_rescript_file_locations.insert(res_file_path);
-                                    }
-                                    None => (),
-                                }
-                            }
-                            _ => (),
-                        },
-                        None => (),
-                    }
-                }
-                Err(_) => (),
-            }
-        }
-    }
-
-    // delete the .mjs file which appear in our previous compile assets
-    // but does not exists anymore
-    // delete the compiler assets for which modules we can't find a rescript file
-    // location of rescript file is in the AST
-    // delete the .mjs file for which we DO have a compiler asset, but don't have a
-    // rescript file anymore (path is found in the .ast file)
-    let diff = ast_rescript_file_locations
-        .difference(&rescript_file_locations)
-        .collect::<Vec<&String>>();
-
-    let diff_len = diff.len();
-
-    diff.par_iter().for_each(|res_file_location| {
-        let _ = std::fs::remove_file(helpers::change_extension(res_file_location, "mjs"));
-        let (_module_name, package_name, package_namespace, _last_modified, _ast_file_path) =
-            ast_modules
-                .get(&res_file_location.to_string())
-                .expect("Could not find module name for ast file");
-        remove_asts(
-            res_file_location,
-            package_name,
-            package_namespace,
-            root_path,
-        );
-        remove_compile_assets(
-            res_file_location,
-            package_name,
-            package_namespace,
-            root_path,
-        );
-    });
-
-    ast_rescript_file_locations
-        .intersection(&rescript_file_locations)
-        .into_iter()
-        .for_each(|res_file_location| {
-            let (module_name, _package_name, _package_namespace, ast_last_modified, ast_file_path) =
-                ast_modules
-                    .get(res_file_location)
-                    .expect("Could not find module name for ast file");
-            let module = all_modules
-                .get_mut(module_name)
-                .expect("Could not find module for ast file");
-
-            match &mut module.source_type {
-                SourceType::MlMap(_) => unreachable!("MlMap is not matched with a ReScript file"),
-                SourceType::SourceFile(source_file) => {
-                    if helpers::is_interface_ast_file(ast_file_path) {
-                        let interface = source_file
-                            .interface
-                            .as_mut()
-                            .expect("Could not find interface for module");
-
-                        let source_last_modified = interface.last_modified;
-                        if ast_last_modified > &source_last_modified {
-                            interface.dirty = false;
-                        }
-                    } else {
-                        let implementation = &mut source_file.implementation;
-                        let source_last_modified = implementation.last_modified;
-                        if ast_last_modified > &source_last_modified {
-                            implementation.dirty = false;
-                        }
-                    }
-                }
-            }
-        });
-
-    let ast_module_names = ast_modules
-        .values()
-        .map(|(module_name, _, _, _, _)| module_name)
-        .collect::<AHashSet<&String>>();
-
-    let all_module_names = all_modules
-        .keys()
-        .map(|module_name| module_name)
-        .collect::<AHashSet<&String>>();
-
-    let deleted_module_names = ast_module_names
-        .difference(&all_module_names)
-        .map(|module_name| {
-            // if the module is a namespace, we need to mark the whole namespace as dirty when a module has been deleted
-            if let Some(namespace) = helpers::get_namespace_from_module_name(module_name) {
-                return namespace;
-            }
-            return module_name.to_string();
-        })
-        .collect::<AHashSet<String>>();
-
-    (
-        diff_len,
-        ast_rescript_file_locations.len(),
-        deleted_module_names,
-    )
 }
 
 pub fn get_version(project_root: &str) -> String {
@@ -462,7 +165,7 @@ fn get_dep_modules(
     valid_modules: &AHashSet<String>,
 ) -> AHashSet<String> {
     let mut deps = Vec::new();
-    if let Ok(lines) = read_lines(ast_file.to_string()) {
+    if let Ok(lines) = helpers::read_lines(ast_file.to_string()) {
         // we skip the first line with is some null characters
         // the following lines in the AST are the dependency modules
         // we stop when we hit a line that starts with a "/", this is the path of the file.
@@ -1052,7 +755,7 @@ pub fn compile_file(
                 .to_string();
 
             let dir = std::path::Path::new(implementation_file_path)
-                .strip_prefix(get_package_path(root_path, &module.package.name))
+                .strip_prefix(helpers::get_package_path(root_path, &module.package.name))
                 .unwrap()
                 .parent()
                 .unwrap();
@@ -1061,28 +764,40 @@ pub fn compile_file(
             if !is_interface {
                 let _ = std::fs::copy(
                     build_path_abs.to_string() + "/" + &module_name + ".cmi",
-                    std::path::Path::new(&get_bs_build_path(root_path, &module.package.name))
-                        .join(dir)
-                        .join(module_name.to_owned() + ".cmi"),
+                    std::path::Path::new(&helpers::get_bs_build_path(
+                        root_path,
+                        &module.package.name,
+                    ))
+                    .join(dir)
+                    .join(module_name.to_owned() + ".cmi"),
                 );
                 let _ = std::fs::copy(
                     build_path_abs.to_string() + "/" + &module_name + ".cmj",
-                    std::path::Path::new(&get_bs_build_path(root_path, &module.package.name))
-                        .join(dir)
-                        .join(module_name.to_owned() + ".cmj"),
+                    std::path::Path::new(&helpers::get_bs_build_path(
+                        root_path,
+                        &module.package.name,
+                    ))
+                    .join(dir)
+                    .join(module_name.to_owned() + ".cmj"),
                 );
                 let _ = std::fs::copy(
                     build_path_abs.to_string() + "/" + &module_name + ".cmt",
-                    std::path::Path::new(&get_bs_build_path(root_path, &module.package.name))
-                        .join(dir)
-                        .join(module_name.to_owned() + ".cmt"),
+                    std::path::Path::new(&helpers::get_bs_build_path(
+                        root_path,
+                        &module.package.name,
+                    ))
+                    .join(dir)
+                    .join(module_name.to_owned() + ".cmt"),
                 );
             } else {
                 let _ = std::fs::copy(
                     build_path_abs.to_string() + "/" + &module_name + ".cmti",
-                    std::path::Path::new(&get_bs_build_path(root_path, &module.package.name))
-                        .join(dir)
-                        .join(module_name.to_owned() + ".cmti"),
+                    std::path::Path::new(&helpers::get_bs_build_path(
+                        root_path,
+                        &module.package.name,
+                    ))
+                    .join(dir)
+                    .join(module_name.to_owned() + ".cmti"),
                 );
             }
 
@@ -1117,44 +832,6 @@ pub fn clean(path: &str) {
     })
 }
 
-fn failed_to_compile(module: &Module) -> bool {
-    match &module.source_type {
-        SourceType::SourceFile(SourceFile {
-            implementation:
-                Implementation {
-                    compile_state: CompileState::Error | CompileState::Warning,
-                    ..
-                },
-            ..
-        }) => true,
-        SourceType::SourceFile(SourceFile {
-            interface:
-                Some(Interface {
-                    compile_state: CompileState::Error | CompileState::Warning,
-                    ..
-                }),
-            ..
-        }) => true,
-        SourceType::SourceFile(SourceFile {
-            implementation:
-                Implementation {
-                    parse_state: ParseState::ParseError | ParseState::Warning,
-                    ..
-                },
-            ..
-        }) => true,
-        SourceType::SourceFile(SourceFile {
-            interface:
-                Some(Interface {
-                    parse_state: ParseState::ParseError | ParseState::Warning,
-                    ..
-                }),
-            ..
-        }) => true,
-        _ => false,
-    }
-}
-
 fn is_dirty(module: &Module) -> bool {
     match module.source_type {
         SourceType::SourceFile(SourceFile {
@@ -1168,42 +845,6 @@ fn is_dirty(module: &Module) -> bool {
         SourceType::SourceFile(_) => false,
         SourceType::MlMap(MlMap { dirty, .. }) => dirty,
     }
-}
-
-fn cleanup_after_build(
-    modules: &AHashMap<String, Module>,
-    compiled_modules: &AHashSet<String>,
-    all_modules: &AHashSet<String>,
-    project_root: &str,
-) {
-    let failed_modules = all_modules
-        .difference(&compiled_modules)
-        .collect::<AHashSet<&String>>();
-
-    modules.par_iter().for_each(|(module_name, module)| {
-        if failed_to_compile(module) || failed_modules.contains(module_name) {
-            // only retain ast file if it compiled successfully, that's the only thing we check
-            // if we see a AST file, we assume it compiled successfully, so we also need to clean
-            // up the AST file if compile is not successful
-            match &module.source_type {
-                SourceType::SourceFile(source_file) => {
-                    remove_asts(
-                        &source_file.implementation.path,
-                        &module.package.name,
-                        &module.package.namespace,
-                        &project_root,
-                    );
-                    // remove_compile_assets(
-                    //     &source_file.implementation.path,
-                    //     &module.package.name,
-                    //     &module.package.namespace,
-                    //     &project_root,
-                    // );
-                }
-                _ => (),
-            }
-        }
-    });
 }
 
 pub fn build(path: &str) -> Result<AHashMap<std::string::String, Module>, ()> {
@@ -1252,7 +893,7 @@ pub fn build(path: &str) -> Result<AHashMap<std::string::String, Module>, ()> {
     );
     let timing_cleanup = Instant::now();
     let (diff_cleanup, total_cleanup, deleted_module_names) =
-        cleanup_previous_build(&packages, &mut modules, &project_root);
+        clean::cleanup_previous_build(&packages, &mut modules, &project_root);
     let timing_cleanup_elapsed = timing_cleanup.elapsed();
     println!(
         "{}\r{} {}Cleaned {}/{} {:.2}s",
@@ -1301,7 +942,7 @@ pub fn build(path: &str) -> Result<AHashMap<std::string::String, Module>, ()> {
                 timing_ast_elapsed.as_secs_f64()
             );
             println!("{}", &err);
-            cleanup_after_build(&modules, &AHashSet::new(), &all_modules, &project_root);
+            clean::cleanup_after_build(&modules, &AHashSet::new(), &all_modules, &project_root);
             return Err(());
         }
     }
@@ -1473,7 +1114,7 @@ pub fn build(path: &str) -> Result<AHashMap<std::string::String, Module>, ()> {
     let compile_duration = start_compiling.elapsed();
 
     pb.finish();
-    cleanup_after_build(&modules, &compiled_modules, &all_modules, &project_root);
+    clean::cleanup_after_build(&modules, &compiled_modules, &all_modules, &project_root);
     if compile_errors.len() > 0 {
         println!(
             "{}\r{} {}Compiled in {:.2}s",
